📦 refactor: validation for not found entities

+++ src/modules/tank/services/tank/tank.service.ts
@@ -1,9 +1,10 @@
 import { CreateTankDto, UpdateTankDto } from 'src/modules/tank/dto/tank.dto';
 import { Tank } from 'src/modules/tank/models/tank.entity';
 import { TankRepository } from 'src/modules/tank/tank.repository';
+import { eMessages } from 'src/shared/enums/messages.enum';
 import { DeleteResult, UpdateResult } from 'typeorm';
 
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 
 @Injectable()
@@ -21,8 +22,12 @@ export class TankService {
     });
   }
 
-  getTankById(id: number): Promise<Tank> {
-    return this.tankRepository.findOne(id);
+  async getTankById(id: number): Promise<Tank> {
+    const tank = await this.tankRepository.findOne(id);
+    if (!tank)
+      throw new HttpException(eMessages.NOT_FOUND, HttpStatus.NOT_FOUND);
+
+    return tank;
   }
 
   createTank(tank: CreateTankDto): Promise<Tank> {
@@ -33,11 +38,19 @@ export class TankService {
     return this.tankRepository.save(newTank);
   }
 
-  updateTank(id: number, tank: UpdateTankDto): Promise<UpdateResult> {
+  async updateTank(id: number, tank: UpdateTankDto): Promise<UpdateResult> {
+    const targetTank = await this.tankRepository.findOne(id);
+    if (!targetTank)
+      throw new HttpException(eMessages.NOT_FOUND, HttpStatus.NOT_FOUND);
+
     return this.tankRepository.update(id, tank);
   }
 
-  deleteTank(id: number): Promise<DeleteResult> {
+  async deleteTank(id: number): Promise<DeleteResult> {
+    const targetTank = await this.tankRepository.findOne(id);
+    if (!targetTank)
+      throw new HttpException(eMessages.NOT_FOUND, HttpStatus.NOT_FOUND);
+
     return this.tankRepository.delete(id);
   }
